refactor putNesInElasticSearch.py with py modules - to be tested
