+++ modules/configParser_functs.py
@@ -2,14 +2,14 @@
 # Fuctions to access config parser
 # Author Lorenzo Allori <lorenzo.allori@gmail.com>
 
-
 from configparser import ConfigParser
 import os
+import sys
 
-def read_configfile():
+configFile="configFile.ini"
+configFileMissing="Config file not found - maybe need to edit configFile_default.ini and rename it to configFile.ini?"
 
-    configFile="configFile.ini"
-
+def mysql_properties():
     # Get config properties
     myhost=str()
     myport=str()
@@ -26,5 +26,19 @@ def read_configfile():
         mypasswd=parser.get('MYSQLDATABASE', 'mysql.passwd')
         mydbname=parser.get('MYSQLDATABASE', 'mysql.dbname')
     else:
-        print("Config file not found")
+        print(configFileMissing)
+        sys.exit()
     return myhost, myport, myuser, mypasswd, mydbname
+
+def elastic_properties():
+    if os.path.isfile(configFile):
+        parser = ConfigParser()
+        parser.read(configFile)
+        eshost=parser.get('ELASTICSEARCH', 'elasticsearch.host')
+        esport=parser.get('ELASTICSEARCH', 'elasticsearch.port')
+        esuser=parser.get('ELASTICSEARCH', 'elasticsearch.username')
+        espassword=parser.get('ELASTICSEARCH', 'elasticsearch.password')
+    else:
+        print(configFileMissing)
+        sys.exit()
+    return eshost, esport, esuser, espassword
