+++ modules/utils_funct.py
@@ -10,3 +10,11 @@ def currentDateAndTime():
     current_date = now.strftime("%Y-%m-%d")
     current_time = now.strftime("%H:%M:%S")
     return current_date, current_time
+
+
+def filldate(s):
+    d='/'.join(x.zfill(2) for x in d.split('/'))
+    return s
+
+def existstr(s):
+    return '' if s is None else str(s)
